requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
idna==2.10
multidict==4.7.6
Pillow==7.2.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
selenium==3.141.0
urllib3==1.25.9
webdriver-manager==3.2.1
yarl==1.4.2
